Move ViewOnlyTag to it's own file
