+++ frontend/src/metabase/query_builder/components/view/ViewHeader/components/SavedQuestionLeftSide/SavedQuestionLeftSide.tsx
@@ -1,6 +1,5 @@
 import type React from "react";
 import { useCallback, useEffect, useState } from "react";
-import { t } from "ttag";
 
 import SavedQuestionHeaderButton from "metabase/query_builder/components/SavedQuestionHeaderButton/SavedQuestionHeaderButton";
 import {
@@ -12,16 +11,14 @@ import {
   ViewHeaderLeftSubHeading,
   ViewHeaderMainLeftContentContainer,
 } from "metabase/query_builder/components/view/ViewHeader/ViewTitleHeader.styled";
-import { Flex, HoverCard, Icon, Text, rem } from "metabase/ui";
-import * as Lib from "metabase-lib";
+import { Flex } from "metabase/ui";
 import type Question from "metabase-lib/v1/Question";
 
 import { HeadBreadcrumbs } from "../HeaderBreadcrumbs";
 import { HeaderCollectionBadge } from "../HeaderCollectionBadge";
 import { QuestionDataSource } from "../QuestionDataSource";
 
-import CS from "./SavedQuestionLeftSide.module.css";
-import { useHiddenSourceTables } from "./hooks";
+import { ViewOnlyTag } from "./ViewOnly";
 
 interface SavedQuestionLeftSideProps {
   question: Question;
@@ -95,7 +92,7 @@ export function SavedQuestionLeftSide({
               ]}
             />
 
-            <ReadOnlyTag question={question} />
+            <ViewOnlyTag question={question} />
           </Flex>
         </SavedQuestionHeaderButtonContainer>
       </ViewHeaderMainLeftContentContainer>
@@ -119,35 +116,5 @@ export function SavedQuestionLeftSide({
         </ViewHeaderLeftSubHeading>
       )}
     </SavedQuestionLeftSideRoot>
-  );
-}
-
-export function ReadOnlyTag({ question }: { question: Question }) {
-  const { isEditable } = Lib.queryDisplayInfo(question.query());
-  const hiddenSourceTables = useHiddenSourceTables(question);
-
-  if (isEditable) {
-    return null;
-  }
-
-  const tableName = hiddenSourceTables[0]?.displayName;
-
-  return (
-    <HoverCard position="bottom-start" disabled={!tableName}>
-      <HoverCard.Target>
-        <Flex align="center" gap="xs" px={4} py={2} mt={4} className={CS.badge}>
-          <Icon name="lock_filled" size={12} />
-          <Text size="xs" fw="bold">
-            {t`View-only`}
-          </Text>
-        </Flex>
-      </HoverCard.Target>
-      <HoverCard.Dropdown>
-        <Text
-          maw={rem(360)}
-          p="md"
-        >{t`One of the administrators hid the source table “${tableName}”, making this question view-only.`}</Text>
-      </HoverCard.Dropdown>
-    </HoverCard>
   );
 }
